+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 # 0.3.0 (29-Oct-15)
 - forward published/consume object validation method to the current service in the infrastructure
 - forward wrapped consume object creation to the current service
+- use HTTP_USER and HTTP_PASSWORD instead of GITHUB_USER and GITHUB_TOKEN
 
 # 0.2.5 (28-Oct-15)
 - better CLI help